Move the command line project to CLI namespace
