--- src/XHarness/Properties/launchSettings.json
@@ -1,8 +1,0 @@
-{
-  "profiles": {
-    "XHarness": {
-      "commandName": "Project",
-      "commandLineArgs": " ios package -n=FoobittyBar --working-directory=workdir --output-directory=outdir"
-    }
-  }
-}
