+++ src/XHarness/XHarness.sln
@@ -1,7 +1,8 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XHarness", "XHarness.csproj", "{824A7320-E35F-40D9-8CC5-6FC16383DADF}"
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29920.165
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.XHarness.CLI", "Microsoft.DotNet.XHarness.CLI\Microsoft.DotNet.XHarness.CLI.csproj", "{A2966C56-1F72-44BF-8916-4D981E193F57}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -9,9 +10,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{824A7320-E35F-40D9-8CC5-6FC16383DADF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{824A7320-E35F-40D9-8CC5-6FC16383DADF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{824A7320-E35F-40D9-8CC5-6FC16383DADF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{824A7320-E35F-40D9-8CC5-6FC16383DADF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A2966C56-1F72-44BF-8916-4D981E193F57}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A2966C56-1F72-44BF-8916-4D981E193F57}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A2966C56-1F72-44BF-8916-4D981E193F57}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A2966C56-1F72-44BF-8916-4D981E193F57}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F89AD1BD-BD08-41EC-A45F-7C7C10D63055}
 	EndGlobalSection
 EndGlobal
